:recycle: 링크 상세 화면 전용 ViewModel 분리
